+++ v2/transport/http/server/server.go
@@ -61,7 +61,7 @@ func NewServer(
 		s.errorEncoder = DefaultErrorEncoder
 	}
 	if s.errorHandler == nil {
-		s.errorHandler = transport.NewLogErrorHandler(nil)
+		s.errorHandler = transport.NopErrorHandler
 	}
 	return s
 }